+++ src/alias/initial.sh
@@ -2,9 +2,10 @@
 
 _myzs:internal:module:initial "$0"
 
-if $MYZS_SETTINGS_AUTOLOAD_SETUP_LOCAL; then
+if _myzs:internal:setting:is-enabled "setup-file/automatic"; then
   cd_function() {
     cd "$1" && myzs-setup-local
   }
+
   _myzs:internal:alias-force "cd" "cd_function"
 fi