+++ src/app/fzf.sh
@@ -1,5 +1,0 @@
-# shellcheck disable=SC1090,SC2148
-
-_myzs:internal:module:initial "$0"
-
-_myzs:internal:load "fezzy search script" "$HOME/.fzf.zsh"