+++ src/utils/helper/checker.sh
@@ -76,11 +76,11 @@ _myzs:internal:checker:string-exist() {
 }
 
 _myzs:internal:checker:small-type() {
-  myzs:setting:is "myzs/type" "SMALL"
+  _myzs:internal:setting:is "myzs/type" "SMALL"
 }
 
 _myzs:internal:checker:fully-type() {
-  myzs:setting:is "myzs/type" "FULLY"
+  _myzs:internal:setting:is "myzs/type" "FULLY"
 }
 
 _myzs:internal:checker:mac() {