+++ src/utils/helper/logger.sh
@@ -21,7 +21,7 @@ _myzs:private:log() {
   local logger_level="$1"
   shift
 
-  if echo "${MYZS_LOG_LEVEL[*]}" | grep -iqF "$logger_level"; then
+  if _myzs:internal:setting:contains "logger/level" "$logger_level"; then
     if [[ "$__MYZS__LOGDIR" != "" ]] && [[ $__MYZS__LOGFILE != "" ]]; then
       if ! test -d "$__MYZS__LOGDIR"; then
         mkdir -p "$__MYZS__LOGDIR" >/dev/null
@@ -61,42 +61,4 @@ _myzs:internal:log:warn() {
 
 _myzs:internal:log:error() {
   _myzs:private:log "ERROR" "$@"
-}
-
-_myzs:internal:log:set-level() {
-  _myzs:internal:log:info "update log level to $*"
-
-  export MYZS_PREVIOUS_LOG_LEVEL=("${MYZS_LOG_LEVEL[@]}")
-
-  unset MYZS_LOG_LEVEL
-  export MYZS_LOG_LEVEL=("$@")
-}
-
-_myzs:log:silent() {
-  _myzs:internal:log:info "update log level to silent mode"
-
-  export MYZS_PREVIOUS_LOG_LEVEL=("${MYZS_LOG_LEVEL[@]}")
-
-  unset MYZS_LOG_LEVEL
-  export MYZS_LOG_LEVEL=()
-}
-
-_myzs:log:debug() {
-  _myzs:internal:log:set-level "error" "warn" "info" "debug"
-}
-
-_myzs:log:info() {
-  _myzs:internal:log:set-level "error" "warn" "info"
-}
-
-_myzs:log:warn() {
-  _myzs:internal:log:set-level "error" "warn"
-}
-
-_myzs:log:error() {
-  _myzs:internal:log:set-level "error"
-}
-
-_myzs:log:revert() {
-  _myzs:internal:log:set-level "${MYZS_PREVIOUS_LOG_LEVEL[@]}"
 }