+++ src/utils/helper/setting.sh
@@ -3,18 +3,16 @@
 # 1. String | Number
 #    - Data storage: "${__MYZS__SETTING_PREFIX}${NAME}=${VALUE}"
 #                    e.g. __MYZS_SETTINGS__MODULE_NAME="setting"
-#    - Data access:  using `myzs:setting:is "${NAME}" "${VALUE}"`
-#                    e.g. myzs:setting:is "module/name" "setting"
+#    - Data access:  using `_myzs:internal:setting:is "${NAME}" "${VALUE}"`
+#                    e.g. _myzs:internal:setting:is "module/name" "setting"
 # 2. Boolean
 #    - Data storage: using myzs:settings:loading(loader enabled disabled) enabled and disabled method
 #                    e.g. enabled "module-loader"
 
-_myzs:internal:module:initial "$0"
-
 export __MYZS__SETTING_PREFIX="__MYZS_SETTINGS__"
 
 # All boolean settings will prefix with _ENABLED / _DISABLED and value will always be true
-# e.g. myzs:setting:is-enabled "analytics" will try to query from $__MYZS_SETTINGS__ANALYTICS_ENABLED must be 'true'
+# e.g. _myzs:internal:setting:is-enabled "analytics" will try to query from $__MYZS_SETTINGS__ANALYTICS_ENABLED must be 'true'
 
 _myzs:private:setting:variable() {
   local name="$1"
@@ -23,6 +21,8 @@ _myzs:private:setting:variable() {
 }
 
 _myzs:private:setting:setup() {
+  local variable key value
+
   key="$(_myzs:private:setting:variable "$1")"
   value="$2"
 
@@ -36,6 +36,24 @@ _myzs:private:setting:enabled() {
 
 _myzs:private:setting:disabled() {
   _myzs:private:setting:setup "$1" "false"
+}
+
+_myzs:private:setting:array() {
+  local key value="" variable
+
+  key="$(_myzs:private:setting:variable "$1")"
+  shift 1
+
+  for data in "$@"; do
+    if test -z "$value"; then
+      value="\"$data\""
+    else
+      value="$value \"$data\""
+    fi
+  done
+
+  variable="${__MYZS__SETTING_PREFIX}${key}"
+  eval "$variable=($value)"
 }
 
 _myzs:private:setting:checker() {
@@ -79,20 +97,20 @@ _myzs:internal:setting:initial() {
 }
 
 # use this when default value is disabled (false)
-myzs:setting:is-enabled() {
-  myzs:setting:is "$1" "true"
+_myzs:internal:setting:is-enabled() {
+  _myzs:internal:setting:is "$1" "true"
 }
 
 # use this when default value is enabled (true)
-myzs:setting:is-disabled() {
-  myzs:setting:is "$1" "false"
+_myzs:internal:setting:is-disabled() {
+  _myzs:internal:setting:is "$1" "false"
 }
 
 _myzs:private:setting:is() {
   local data="$1" checker="$2"
   [[ "$data" == "$checker" ]]
 }
-myzs:setting:is() {
+_myzs:internal:setting:is() {
   _myzs:private:setting:checker "_myzs:private:setting:is" "$@"
 }
 
@@ -101,7 +119,7 @@ _myzs:private:setting:greater-than() {
   [[ "$data" -gt "$checker" ]]
 }
 # $setting > $B && $setting > $C ...
-myzs:setting:greater-than() {
+_myzs:internal:setting:greater-than() {
   _myzs:private:setting:checker "_myzs:private:setting:greater-than" "$@"
 }
 
@@ -110,8 +128,17 @@ _myzs:private:setting:less-than() {
   [[ "$data" -lt "$checker" ]]
 }
 # $setting < $B && $setting < $C ...
-myzs:setting:less-than() {
+_myzs:internal:setting:less-than() {
   _myzs:private:setting:checker "_myzs:private:setting:less-than" "$@"
+}
+
+_myzs:private:setting:contains() {
+  local data="$1" checker="$2"
+  echo "${data}" | grep -iqF "$checker"
+}
+# 'a b c' contains 'a' && 'a b c' contains 'b'
+_myzs:internal:setting:contains() {
+  _myzs:private:setting:checker "_myzs:private:setting:contains" "$@"
 }
 
 myzs:setting:get() {
@@ -121,4 +148,11 @@ myzs:setting:get() {
   eval "__vtest=\"\${${variable}:-$fallback}\""
 
   echo "$__vtest"
+}
+
+myzs:setting:get-array() {
+  local name="$1" varname="$2" variable __vtest=""
+
+  variable="${__MYZS__SETTING_PREFIX}$(_myzs:private:setting:variable "${name}")"
+  eval "$varname=(\${${variable}})"
 }